Fix template deployment script (Windows/PowerShell) (#84)
* dev: Add the README.md file to the project to find easily

* dev: Styling update for color reference

* dev: Fixup template project and Windows register script

Fixes based on issues encountered on evening stream of Jan 29, 2019

* bugfix/Fix-template-deployment: Reset formatting from DEV branch

* bugfix/Fix-template-deployment: Re-add README to template project file

* bugfix/Fix-template-deployment: Copy changes from the template PowerShell script to the sample plugin

+++ src/SamplePlugin/RegisterPluginAndStartStreamDeck.ps1
@@ -1,22 +1,58 @@
-﻿
-$streamDeckExePath = "C:\Program Files\Elgato\StreamDeck\StreamDeck.exe"
-$bindir = $PSScriptRoot + "\bin\Debug\netcoreapp2.2\win-x64"
-$manifestFile = $bindir +"\manifest.json"
+﻿Write-Host "Gathering deployment items..."
 
+Write-Host "Script root: $PSScriptRoot`n"
+
+$basePath = $PSScriptRoot
+
+if ($PSSCriptRoot.Length -eq 0) {
+  $basePath = $PWD.Path;
+}
+
+
+# Load and parse the plugin project file
+$pluginProjectFile = "$basePath\_StreamDeckPlugin_.csproj"
+$projectContent = Get-Content $pluginProjectFile | Out-String;
+$projectXML = [xml]$projectContent;
+
+$buildConfiguration = "Debug"
+
+# Get the target .net core framework
+$targetFrameworkName = $projectXML.Project.PropertyGroup.TargetFramework;
+
+# Set local path references
+$streamDeckExePath = "$($ENV:ProgramFiles)\Elgato\StreamDeck\StreamDeck.exe"
+
+# For now, this PS script will only be run on Windows.
+$bindir = "$basePath\bin\Debug\$targetFrameworkName\win-x64\"
+
+# Make sure we actually have a directory/build to deploy
+If (-not (Test-Path $bindir)) {
+  Write-Error "The output directory `"$bindir`" was not found.`n You must first build the `"_StreamDeckPlugin_`" project before calling this script.";
+  exit 1;
+}
+
+# Load and parse the plugin's manifest file
+$manifestFile = $bindir +"\manifest.json"
 $manifestContent = Get-Content $manifestFile | Out-String
 $json = ConvertFrom-JSON $manifestcontent
+
 $uuidAction = $json.Actions[0].UUID
+
 $pluginID = $uuidAction.substring(0, $uuidAction.Length - ".action".Length)
-$destDir = $env:APPDATA + "\Elgato\StreamDeck\Plugins\" + $pluginID + ".sdPlugin"
+$bindir = $basePath + "\bin\$buildConfiguration\$targetFrameworkName"
+$destDir = "$($env:APPDATA)\Elgato\StreamDeck\Plugins\$pluginID.sdPlugin"
 
-$pluginName = split-path $PSScriptRoot -leaf
+$pluginName = Split-Path $basePath -leaf
 
-Get-Process StreamDeck,$pluginName | Stop-Process –force -ErrorAction SilentlyContinue 
+Get-Process StreamDeck,$pluginName | Stop-Process –force -ErrorAction SilentlyContinue
 
-mkdir $destDir -ErrorAction SilentlyContinue
-Get-ChildItem -Path $bindir -Recurse | ? { $_.FullName -inotmatch '.vs' } | Copy-Item -Destination $destDir
+# Delete the target directory, make sure the deployment/copy is clean
+Remove-Item -Recurse -Path $destDir
 
-Write-Host "Script Completed"
+# Then copy all deployment items to the plugin directory
+New-Item -Type Directory -Path $destDir -ErrorAction SilentlyContinue | Out-Null
+Copy-Item -Path $bindir -Destination $destDir -Recurse
 
+Write-Host "Deployment complete. Restarting Stream Deck..."
 Start-Process $streamDeckExePath
 exit 0
